Bump version to 4.0.14-SNAPSHOT and update deprecated delete method version
Update `pom.xml` to set the version to `4.0.14-SNAPSHOT` for development. Adjust the deprecation version of the `delete(String path, boolean recursive)` method in `KeelFileHelper` from `4.0.13` to `4.0.14` to align with the new versioning. Update the `@since` tags for `delete(String path)` and `deleteRecursive(String path)` methods to reflect the correct version `4.0.14`.

Signed-off-by: sinri <e.joshua.s.e@gmail.com>